+++ src/main/java/name/ulbricht/streams/impl/intermediate/Skip.java
@@ -6,27 +6,27 @@ import name.ulbricht.streams.api.Configuration;
 import name.ulbricht.streams.api.ConfigurationType;
 import name.ulbricht.streams.api.IntermediateOperation;
 
-@Configuration(name = "skip", type = ConfigurationType.INTEGER, displayName = "Skip Elements")
+@Configuration(name = "skip", type = ConfigurationType.LONG, displayName = "Skip Elements")
 public final class Skip implements IntermediateOperation<Object, Object> {
 
-	private int skip = 5;
+	private long skip = 5;
 
-	public int getSkip() {
+	public long getSkip() {
 		return this.skip;
 	}
 
-	public void setSkip(final int skip) {
+	public void setSkip(final long skip) {
 		this.skip = skip;
 	}
 
 	@Override
 	public String getSourceCode() {
-		return String.format(".skip(%s)", Integer.toString(this.skip));
+		return String.format(".skip(%s)", Long.toString(this.skip));
 	}
 
 	@Override
 	public String getConfigurationText() {
-		return String.format("skip=%s", Integer.toString(this.skip));
+		return String.format("skip=%s", Long.toString(this.skip));
 	}
 
 	@Override
